Performance enhancements and diminutives
I found a way to speed up performance, but it requires converting all of the yob*.txt files to a format similar to the surname files. Only the yob1972.txt file has been converted as part of this commit.

Added functionality for diminutives for names that typically have them (e.g. David, Charles, Peter, etc.). Where a diminutive for a first name exists:

  1. There is a 50% chance it will be used
  2. If it is used, the diminutive for the name is randomly selected from
     the available list